net: ipv6: Check on-link prefix before falling back to default router (DO NOT MERGE yet)
When routing packets, check if the destination address matches an
on-link prefix on any interface before using the default router.
This ensures packets destined for on-link prefixes are not
incorrectly sent to the default router.
Note: this is a draft implementation and should not be merged yet
before further testing. Anyways, this enables that no neighbors should be
needed in dect nr+ sink, i.e. routing from Internet to DECT NR+
based on set prefix.

Signed-off-by: Jani Hirsimäki <jani.hirsimaki@nordicsemi.no>